// projects/webapp/src/app/features/contract-management/constants/transfer-state.enum.ts
export enum TransferStateEnum {
  INITIAL = 'INITIALIZED',
  PROVISIONING = 'PROVISIONING',
  PROVISIONING_REQUESTED = 'PROVISIONING REQUESTED',
  PROVISIONED = 'PROVISIONED',
  REQUESTING = 'REQUESTING',
  REQUESTED = 'REQUESTED',
  STARTING = 'STARTING',
  STARTED = 'STARTED',
  SUSPENDING = 'SUSPENDING',
  SUSPENDING_REQUESTED = 'SUSPENDING REQUESTED',
  SUSPENDED = 'SUSPENDED',
  RESUMING = 'RESUMING',
  RESUMED = 'RESUMED',
  COMPLETING = 'COMPLETING',
  COMPLETING_REQUESTED = 'COMPLETING REQUESTED',
  COMPLETED = 'COMPLETED',
  TERMINATING = 'TERMINATING',
  TERMINATING_REQUESTED = 'TERMINATING REQUESTED',
  TERMINATED = 'TERMINATED',
  DEPROVISIONING = 'DEPROVISIONING',
  DEPROVISIONING_REQUESTED = 'DEPROVISIONING REQUESTED',
  DEPROVISIONED = 'DEPROVISIONED'
}